backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Firebase dependencies
firebase-admin==6.4.0
google-cloud-firestore==2.13.1

# AI/ML dependencies
openai==1.6.1
numpy==1.26.2

# Social Media API dependencies
tweepy==4.14.0
facebook-sdk==3.1.0
requests==2.31.0

# Data processing dependencies
python-dotenv==1.0.0

# Utility dependencies
python-dateutil==2.8.2

# Google Analytics & Ads Integration
google-analytics-data==0.18.4
google-ads==27.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0

# Autonomous processing & scheduling
celery==5.3.4
redis==5.0.1
APScheduler==3.10.4

# Advanced analytics and optimization
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4

# Performance monitoring and rate limiting
ratelimit==2.2.1
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Build dependencies to ensure compatibility
setuptools>=65.0.0
wheel>=0.38.0
